+++ src/pages/User/Signup.tsx
@@ -1,5 +1,5 @@
 import { FormEvent, useState, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 
 import { UserRequests } from "../../shared/utils/requests/User.requests";
 import { FiArrowLeft } from "react-icons/fi";
@@ -8,11 +8,12 @@ import Button from "../../shared/components/Button";
 import Footer from "../../shared/components/Footer";
 import Header from "../../shared/components/Header";
 import TextField from "../../shared/components/TextField";
+import ChoiceField from "../../shared/components/ChoiceField";
 
 import SessionController from "../../shared/utils/handlers/SessionController";
-import "../../shared/styles/pages/login/UserEdit.css";
+import "../../shared/styles/pages/user/Signup.css";
 
-export default function UserEdit() {
+export default function Signup() {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [name, setName] = useState("");
@@ -23,6 +24,11 @@ export default function UserEdit() {
 
   const navigate = useNavigate();
 
+  const userProfiles = [
+    { value: "client", label: "Cliente", selected: false },
+    { value: "support", label: "Suporte", selected: false },
+    { value: "admin", label: "Administrador", selected: false },
+  ];
 
   const token = SessionController.getToken();
   const user = SessionController.getUserInfo();
@@ -33,7 +39,7 @@ export default function UserEdit() {
     }
   }, []);
 
-  async function submitUserEdit(event: FormEvent) {
+  async function submitSignup(event: FormEvent) {
     event.preventDefault();
     if (email === "" || password === "" || name === "" || lastname === "") {
       return alert("Preencha todos os campos");
@@ -48,7 +54,7 @@ export default function UserEdit() {
     };
     console.log(payload);
 
-    const response = await userRequest.updateRequest(payload);
+    const response = await userRequest.registerRequest(payload);
 
     if (response?.status === 201) {
       alert("Usuário cadastrado com sucesso!");
@@ -58,11 +64,11 @@ export default function UserEdit() {
   }
 
   return (
-    <div id="userEdit">
-      <div className="userEdit-container">
+    <div id="signup">
+      <div className="signup-container">
         <Header hiddenDropdown={true} />
-        <div className="userEdit-content">
-          <section className="userEdit-welcome">
+        <div className="signup-content">
+          <section className="signup-welcome">
             <h2>
               <FiArrowLeft
                 color="var(--color-withe-main)"
@@ -71,30 +77,30 @@ export default function UserEdit() {
                 }}
               />
             </h2>
-            <h1>Editar perfil</h1>
+            <h1>Cadastre a conta</h1>
           </section>
-          <form className="userEdit-form" onSubmit={submitUserEdit}>
+          <form className="signup-form" onSubmit={submitSignup}>
             <section className="form-sections">
-              <section className="userEdit-data">
+              <section className="signup-data">
                 <label htmlFor="name">Nome</label>
                 <TextField
                   type="text"
-                  placeholder={user?.firstName}
+                  placeholder="John"
                   onChange={(event) => setName(event.target.value)}
                   name="name"
                 />
                 <label htmlFor="lastname">Sobrenome</label>
                 <TextField
                   type="text"
-                  placeholder={user?.lastName}
+                  placeholder="Snow"
                   onChange={(event) => setLastname(event.target.value)}
                   name="lastname"
                 />
               </section>
-              <section className="userEdit-data">
+              <section className="signup-data">
                 <label htmlFor="email">E-mail</label>
                 <TextField
-                  placeholder={user?.email}
+                  placeholder="john.snow@email.com"
                   onChange={(event) => setEmail(event.target.value)}
                   name="email"
                 />
@@ -107,20 +113,19 @@ export default function UserEdit() {
                 />
               </section>
             </section>
-            <section className="userEdit-role">
-              <label htmlFor="role">Cargo</label>
-              <TextField
-                onChange={(event) => setRole(event.target.value)}
-                name="profile_type"
-                width="20rem"
-                placeholder={user?.role}
-                disabled={true}
-              />
-            </section>
-            <section className="userEdit-submit">
-              <Button width="15rem" type="submit">
-                Confirmar alteração
-              </Button>
+            <section className="signup-data-line">
+              <section className="signup-data">
+                <ChoiceField
+                  onChange={(event) => setRole(event.target.value)}
+                  name="profile_type"
+                  items={userProfiles}
+                />
+              </section>
+              <section className="signup-data">
+                <Button width="100%" type="submit">
+                  Cadastrar
+                </Button>
+              </section>
             </section>
           </form>
         </div>
