+++ src/pages/User/UserEdit.tsx
@@ -1,5 +1,5 @@
 import { FormEvent, useState, useEffect } from "react";
-import { Link, useNavigate } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 import { UserRequests } from "../../shared/utils/requests/User.requests";
 import { FiArrowLeft } from "react-icons/fi";
@@ -8,12 +8,11 @@ import Button from "../../shared/components/Button";
 import Footer from "../../shared/components/Footer";
 import Header from "../../shared/components/Header";
 import TextField from "../../shared/components/TextField";
-import ChoiceField from "../../shared/components/ChoiceField";
 
 import SessionController from "../../shared/utils/handlers/SessionController";
-import "../../shared/styles/pages/login/Signup.css";
+import "../../shared/styles/pages/user/UserEdit.css";
 
-export default function Signup() {
+export default function UserEdit() {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [name, setName] = useState("");
@@ -24,11 +23,6 @@ export default function Signup() {
 
   const navigate = useNavigate();
 
-  const userProfiles = [
-    { value: "client", label: "Cliente", selected: false },
-    { value: "support", label: "Suporte", selected: false },
-    { value: "admin", label: "Administrador", selected: false },
-  ];
 
   const token = SessionController.getToken();
   const user = SessionController.getUserInfo();
@@ -39,7 +33,7 @@ export default function Signup() {
     }
   }, []);
 
-  async function submitSignup(event: FormEvent) {
+  async function submitUserEdit(event: FormEvent) {
     event.preventDefault();
     if (email === "" || password === "" || name === "" || lastname === "") {
       return alert("Preencha todos os campos");
@@ -54,7 +48,7 @@ export default function Signup() {
     };
     console.log(payload);
 
-    const response = await userRequest.registerRequest(payload);
+    const response = await userRequest.updateRequest(payload);
 
     if (response?.status === 201) {
       alert("Usuário cadastrado com sucesso!");
@@ -64,11 +58,11 @@ export default function Signup() {
   }
 
   return (
-    <div id="signup">
-      <div className="signup-container">
+    <div id="userEdit">
+      <div className="userEdit-container">
         <Header hiddenDropdown={true} />
-        <div className="signup-content">
-          <section className="signup-welcome">
+        <div className="userEdit-content">
+          <section className="userEdit-welcome">
             <h2>
               <FiArrowLeft
                 color="var(--color-withe-main)"
@@ -77,30 +71,30 @@ export default function Signup() {
                 }}
               />
             </h2>
-            <h1>Cadastre a conta</h1>
+            <h1>Editar perfil</h1>
           </section>
-          <form className="signup-form" onSubmit={submitSignup}>
+          <form className="userEdit-form" onSubmit={submitUserEdit}>
             <section className="form-sections">
-              <section className="signup-data">
+              <section className="userEdit-data">
                 <label htmlFor="name">Nome</label>
                 <TextField
                   type="text"
-                  placeholder="John"
+                  placeholder={user?.firstName}
                   onChange={(event) => setName(event.target.value)}
                   name="name"
                 />
                 <label htmlFor="lastname">Sobrenome</label>
                 <TextField
                   type="text"
-                  placeholder="Snow"
+                  placeholder={user?.lastName}
                   onChange={(event) => setLastname(event.target.value)}
                   name="lastname"
                 />
               </section>
-              <section className="signup-data">
+              <section className="userEdit-data">
                 <label htmlFor="email">E-mail</label>
                 <TextField
-                  placeholder="john.snow@email.com"
+                  placeholder={user?.email}
                   onChange={(event) => setEmail(event.target.value)}
                   name="email"
                 />
@@ -113,19 +107,20 @@ export default function Signup() {
                 />
               </section>
             </section>
-            <section className="signup-data-line">
-              <section className="signup-data">
-                <ChoiceField
-                  onChange={(event) => setRole(event.target.value)}
-                  name="profile_type"
-                  items={userProfiles}
-                />
-              </section>
-              <section className="signup-data">
-                <Button width="100%" type="submit">
-                  Cadastrar
-                </Button>
-              </section>
+            <section className="userEdit-role">
+              <label htmlFor="role">Cargo</label>
+              <TextField
+                onChange={(event) => setRole(event.target.value)}
+                name="profile_type"
+                width="20rem"
+                placeholder={user?.role}
+                disabled={true}
+              />
+            </section>
+            <section className="userEdit-submit">
+              <Button width="15rem" type="submit">
+                Confirmar alteração
+              </Button>
             </section>
           </form>
         </div>
